todo-app/backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10

SQLAlchemy==2.0.15
psycopg2-binary==2.9.7
alembic==1.11.1
flask-jwt-extended==4.7.1
python-dotenv==1.0.0